update them xua xoa

+++ admin/modules/category/delete.php
@@ -11,19 +11,29 @@
         $_SESSION['error'] = "Dữ liệu không tồn tại";
         redirectAdmin("category");
     }
-    
-    $num = $db->delete("category",$id);
-    if($num > 0) 
+
+    // Kiem tra xem danh muc da co san pham
+    $is_product = $db->fetchOne("product"," category_id = $id ");
+    if($is_product == NULL)
     {
-        $_SESSION['success'] = "Xóa thành công";
-          redirectAdmin("category");
-    }
-    else
-     {
-          
-             $_SESSION['error'] = " Xóa thất bại";
+
+        $num = $db->delete("category",$id);
+        if($num > 0) 
+        {
+            $_SESSION['success'] = "Xóa thành công";
               redirectAdmin("category");
         }
+        else
+         {
+                 $_SESSION['error'] = "Xóa thất bại";
+                  redirectAdmin("category");
+            }
+    }
+    else {
+        $_SESSION['error'] = " Danh mục có sản phẩm - không thể xóa !";
+        redirectAdmin("category");
+    }
+    
 
     
 ?>
